refactoring out main work view
